Refactor workshop export buttons
